Supplement 860 data with more complete capacity_mwh information from 860m
This commit pdates 860 with all `capacity_mwh` values found in 860m, not just
the batteries found there but missing from 860. As part of this, it also
simplifies the definition of storage gens in 860m to be any that have
capacity_mwh filled in. This is useful because `capacity_mwh` is often missing
from resources in 860.