Allow users to provide profiles for customized atb technologies
This commit allows users to provide profiles for customized ATB technologies in
the same way that they would for standard ones. Previously, if a user defined a
new technology via `modified_atb_new_gen`, if the `new_technology` name was not
in PowerGenome's built-in `NREL_ATB_TECHNOLOGY_MAP`, `map_nrel_atb_technology()`
would fail to find a match, and the resource would be ignored when looking up
time profiles. Now, instead, in these cases, if a match is not found,
PowerGenome will also look for a matching profile based on the technology name
without using map_nrel_atb_technology(). So users can simply add the new
technology to `modified_atb_new_gen`, add a cluster definition to
`renewables_clusters` with the same technology name, and add files to the
RESOURCE_GROUPS and/or RESOURCE_GROUP_PROFILES directories with that technology
name.